support personal overview widget
