+++ packages/widgets/src/widgets/oh-my-github/personal-overview/createPngThumbnail.ts
@@ -1,0 +1,13 @@
+import { Chart as ChartJs, registerables } from 'chart.js';
+import { data, options } from './chart-options';
+
+ChartJs.register(
+  ...registerables,
+);
+
+export default async function (server: {}, props: {}, ctx: CanvasRenderingContext2D) {
+  new ChartJs(ctx, {
+    data: data,
+    options: { animation: false, ...options },
+  });
+}
