+++ packages/widgets/src/widgets/oh-my-github/personal-overview/index.ts
@@ -1,9 +1,11 @@
 export const Widget = () => import('./Widget');
 
+export const createPngThumbnail = () => import('./createPngThumbnail');
+
 export const preferredSize = {
   width: 800,
   height: 418,
 };
 
 export const category = 'Github Personal';
-export const displayName = 'Overview';
+export const displayName = 'Monthly activity trends';
